+++ packages/snippets/src/ant-design-x/src/components-react/bubble.list.tsx
@@ -1,49 +1,81 @@
+import type { BubbleProps } from '@ant-design/x';
 import type { BubbleListProps } from '@ant-design/x/es/bubble/BubbleList';
 import type { AvatarProps } from 'antd';
-import type { RewriteRolesType, RewriteRoleType } from '../types';
+import type { RewriteBubbleListProps, RewriteBubbleProps, RewriteRolesType, RewriteRoleType } from '../types';
 import { Bubble } from '@ant-design/x';
 import { renderVueCompOrSlot } from '@mixte/snippets/ant-design-x/utils';
 import { assertPlainObject, isFunction } from 'mixte';
 import React from 'react';
 
-interface Props extends Omit<BubbleListProps, 'roles'> {
-  roles?: RewriteRolesType;
-}
+export default function (props: RewriteBubbleListProps) {
+  const {
+    roles,
+    items,
+  } = props;
 
-export default function (props: Props) {
-  const { roles } = props;
+  function transform(props: Pick<RewriteBubbleProps, 'avatar' | 'header' | 'footer' | 'messageRender'>) {
+    const avatar = assertPlainObject<AvatarProps>(props.avatar)
+      ? {
+          ...props.avatar,
+          icon: renderVueCompOrSlot(props.avatar.icon),
+        }
+      : renderVueCompOrSlot(props.avatar);
+    const header = renderVueCompOrSlot(props.header);
+    const footer = renderVueCompOrSlot(props.footer);
+    const messageRender = isFunction(props.messageRender)
+      ? (content: string) => renderVueCompOrSlot(props.messageRender!(content))
+      : undefined;
 
-  const finalRoles = React.useMemo(() => {
+    return {
+      ...(avatar === undefined ? {} : { avatar }),
+      ...(header === undefined ? {} : { header }),
+      ...(footer === undefined ? {} : { footer }),
+      ...(messageRender === undefined ? {} : { messageRender }),
+    };
+  }
+
+  const finalItems = React.useMemo((): BubbleListProps['items'] => {
+    return items?.map((item) => {
+      return {
+        ...item as BubbleProps,
+        ...transform({
+          avatar: item.avatar,
+          header: item.header,
+          footer: item.footer,
+          messageRender: item.messageRender,
+        }),
+      };
+    });
+  }, [items]);
+
+  const finalRoles = React.useMemo((): BubbleListProps['roles'] => {
     if (assertPlainObject<RewriteRolesType>(roles)) {
       return Object.fromEntries(
         Object.entries(roles).map(([key, role]: [string, RewriteRoleType]) => {
           return [
             key,
             {
-              ...role,
-              avatar: assertPlainObject<AvatarProps>(role.avatar)
-                ? {
-                    ...role.avatar,
-                    icon: renderVueCompOrSlot(role.avatar.icon),
-                  }
-                : renderVueCompOrSlot(role.avatar),
-              header: renderVueCompOrSlot(role.header),
-              footer: renderVueCompOrSlot(role.footer),
-              messageRender: isFunction(role.messageRender)
-                ? (content: string) => renderVueCompOrSlot(role.messageRender!(content))
-                : undefined,
+              ...role as BubbleProps,
+              ...transform({
+                avatar: role.avatar,
+                header: role.header,
+                footer: role.footer,
+                messageRender: role.messageRender,
+              }),
             },
           ];
         }),
       );
     }
 
-    return roles;
+    // todo
+    return roles as any;
   }, [roles]);
 
   return (
     <Bubble.List
       {...props}
+      items={finalItems}
       roles={finalRoles}
     />
   );